refactor: restructuring packages - separating process from units
