+++ src/scripts/units/utils/getters.tab.js
@@ -13,3 +13,7 @@ export function getUrl() {
 export function getHostAndPath() {
   return window.location.host + window.location.pathname;
 }
+
+export function isHomepageCheck() {
+  return window.location.href === "https://www.phind.com" || window.location.href === "https://www.phind.com/"
+}
